+++ CHANGELOG.md
@@ -1,11 +1,23 @@
-# Changelog - Keep a changelog
+# Changelog
 
 All notable changes to this project will be documented in this file.
 
-The format is based on [Keep a Changelog](http://keepachangelog.com/)
+The format is based on [Keep a Changelog](http://keepachangelog.com/) 
 and this project adheres to [Semantic Versioning](http://semver.org/).
+
+## [0.2.0] - 2017-12-07
+
+### Changed
+
+- Parser improvements
+- Changed the constructor arguments of Changelog, Change and Release classes
+
+### Removed
+
+- Removed static factories. Use `new` instead.
 
 ## 0.1.0 - 2017-12-07
 
 First version
 
+[0.2.0]: https://github.com/oscarotero/keep-a-changelog/compare/v0.1.0...v0.2.0